refactor: VoteFailureType 추가해서 리팩토링

+++ src/main/kotlin/org/team14/webty/voting/listener/VoteFailureEventListener.kt
@@ -9,6 +9,7 @@ import org.springframework.transaction.event.TransactionPhase
 import org.springframework.transaction.event.TransactionalEventListener
 import org.team14.webty.voting.cache.VoteCacheService
 import org.team14.webty.voting.dto.VoteFailureEvent
+import org.team14.webty.voting.enums.VoteFailureType
 
 @Component
 class VoteFailureEventListener(
@@ -21,7 +22,22 @@ class VoteFailureEventListener(
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     @TransactionalEventListener(phase = TransactionPhase.AFTER_ROLLBACK)
     fun handleVoteFailureEvent(event: VoteFailureEvent) {
-        log.info("투표 처리 실패, 트랜잭션 롤백 similarId : {}, userId : {} ", event.similarId, event.userId)
-        voteCacheService.deleteUserVote(event.similarId, event.userId)
+        log.info(
+            "투표 처리 실패, 트랜잭션 롤백 similarId : {}, userId : {} type: {}",
+            event.similarId,
+            event.userId,
+            event.voteFailureType
+        )
+
+        when (event.voteFailureType) {
+            // 투표 요청 실패 시 redis 유저 투표 내역 삭제
+            VoteFailureType.VOTE_FAILURE -> {
+                voteCacheService.deleteUserVote(event.similarId, event.userId)
+            }
+            // 투표 취소 실패 시 삭제된 투표 내역을 복구
+            VoteFailureType.VOTE_CANCEL_FAILURE -> {
+                voteCacheService.setUserVote(event.similarId, event.userId)
+            }
+        }
     }
 }
